Moved qiapi to resource package
